Properly serialize empty BoolQuery
We used to do this in Search.php, but it was changed to check for null instead of an empty string. This breaks clients relying on the previous behavior to work.

AFAICT all other queries that need this special handling already do it, so this should restore the previous behavior.